Added charts and animations
For private events, a piechart was added to show how many users have checked into the event. For public events, a ticker was added (currently not set up to update periodically).

Animations were added to the event tables and number widgets on the Dashview.

A loading animation was also added to EventTable widgets.