Added scatter plot chart for event checkins
Chart is used to provide a visual way to show users that have checked in. A button was added to show a larger version of the chart within a modal.

An invite doughnut chart was added to show how many users have responded to an invite. The chart is separated into four sections for yes, no, maybe, and unanswered responses.

Changes were made to how props are handled in the NumbersWidget to allow for data to be updated properly.

The PieChartWidget was replaced with the Chart component to allow for more flexibility with how (and where) charts can be rendered.