Added manager widget to EventView
Also made small changes to the EventSearch view to allow for searches to be more configurable.

Changes to the Chart component were made to prevent unnecessary rerenders from being performed.

A handler was added to the form submit on the event feed input to prevent page reloads.

Switched from history.replaceState to history.pushState. A listener was added to prevent the same view from being pushed twice in a row.

Small changes were made to the Modal component to allow it to display different sizes.